requirements.txt
streamlit
pandas
python-docx
matplotlib
pymupdf
openpyxl
xlsxwriter
